Bug 708020: Detect and try to repair recursive stream length references.
Previously this would end up in a recursive call running out of stack.
Now, we avoid marking the stream object as unused if the stream length
indirectly references the stream object itself. Moreover we use the
normal API to open a stream and skip over its contents to determine
the actual stream length.