Add "Fake Bold" detection to stext device.
We add a new FZ_STEXT_COLLECT_FLAGS flag to the device. This is
used to control both the underline/strikethrough detection and
this new "fake bold" detecion.

If set, whenever we add a new character to the page, we first
check to see if the same char has been put in (almost) the same
place before with the same font. (This requires a full page
search, so may potentially slow things down!)

If we find that a char HAS been put in almost the same place
before, we mark that char as being 'fake bold' and then do not
add the new char.

# Conflicts:
#	include/mupdf/fitz/structured-text.h
#	source/fitz/stext-device.c
#	source/tools/mudraw.c