Bug 708079: EPub: Fail more gracefully on missing chapters.
Generate an error page, and set a new error flag in the fz_html
structure. If this flag is set, then when the html page is run
we either mark an error in the cookie, or (in the case where we
have no cookie to mark) we throw an error.

This means that renderers can continue to display the rest of
the file content.

The mupdf viewer already copes with this case.

mupdf-gl is updated here, so that it passes a cookie, and spots
such errors, reporting them in the title bar.

mutool similarly now reports such errors with a warning.

# Conflicts:
#	platform/gl/gl-main.c