crash in `./gperf -h` leads us down the rabbit hole into our abort/exit override code (/source/system_override/), which we now add separate test applications for as our `gperf` is, strangely enough, the ONLY build that's CRASHING on the overridden `exit(0)` call?!?!
Turns out our hack (see common-project.props: `<PreprocessorDefinitions>abort=qiqqa_abort_application;exit=qiqqa_exit_application;...`) does not fly any more, neither in debug mode or release mode with latest MSVC2022: the calls to `exit()` and `abort()` are encoded are indirect calls (`call qword ptr [__imp_abort]`) which is replaced by our *statically linked* override, which DOES NOT provide that indirect jump address slot but straight to the code instead.

So, basically, we'll have to re-hack our abort+exit overrides as we MUST provide that 'indirect call/jump' behaviour the compiler is producing anyway.

Ergo: the current abort/exit override code is still BROKEN in this commit. This commit marks the end of the old and the next one will be a new inject/hook/trampoline implementation that does not segfault like the current one.
One path of investigation is to bring out the heavy guns and use open sourced Microsoft Detours for this... but first, let's see if I can get away with a cheaper home-grown approach first.  ;-)