Bug 701638: Tweak logic for pushing a knockout group in text objects.
Previously, we'd only push a knockout group in case of interaction
between stroke/fill alphas.

Now, we push a knockout group whenever we get a stroke or fill
with a non-"zero or one" alpha, or a strange blend mode.

This is required to handle cases where glyphs overlap (as seen
in the bug examples with shaped text).