Bug703060.pdf: Cache PDF Softmasks as DisplayLists.
The file in the bug sets a softmask, and then draws many many
paths through it.

This means we parse and rerender the softmask multiple times
(once per path).

At the moment, the softmasks are rendered only large enough for
the path bboxes, which makes it hard to reuse the rendered
form.

Rendering every softmask larger would likely not pay off.

So, here, we render the softmask to a display list, and cache that
list. This avoids the reparsing cost.